Some improvements of logLik method
As this is an implementation of the logLik method of base R, and the
base method does not print anything except by the print method of its
return value of class 'logLik', I think this method should also not do
printing on its own. Therefore this commit removes the corresponding
code.

Also, it is unconventional that the logLik method modifies its argument
in-place if the requested value is not present. Therefore, this commit
adds an option to turn this off. The commented out code after the
comment "OR: .." was used to provide the alternative, which is giving a
message and not returning anything.

In the seealso section, links to the functions that can be used to
add these missing likelihoods were added.